% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{Coverage}
\alias{Coverage}
\title{95\% coverage for a Blend object with structural identification}
\usage{
Coverage(x)
}
\arguments{
\item{x}{Blend object.}
}
\value{
coverage
}
\description{
calculate 95\% coverage for varying effects and constant effects under example data
}
\examples{
data(dat)
fit = Blend(y,x,t,J,kn,degree)
Coverage(fit)

}
\seealso{
\code{\link{Blend}}
}
