\name{loss}
\alias{loss}
\title{Loss function}
\usage{
  loss(vector1, vector2)
}
\arguments{
  \item{vector1}{first vector}

  \item{vector2}{second vector}
}
\description{
  Computes the root mean squared difference or percentage
  difference of two numeric or two two-level factors
  respectively.
}
\details{
  The function takes two vectors, which it expects to be of
  the same type, and computes the root mean squared
  difference in the numeric case and the percentage
  difference in the two-level factor case.
}
\author{
  Alex Ustian <alex.l.ustian@gmail.com>
}

