\name{summary.blend}
\alias{summary.blend}
\title{Summary function for blend objects}
\usage{
  summary.blend(blendObj)
}
\arguments{
  \item{blendObj}{blend object for the summary}
}
\description{
  Displays the fitted linear/logistic blended model as well
  as some of the resampling statistics from the
  cross-validation.
}
\author{
  Alex Ustian <alex.l.ustian@gmail.com>
}

