% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLEND.R
\name{BLEND}
\alias{BLEND}
\title{BLEND}
\usage{
BLEND(
  bulk,
  phi,
  alpha = 0.001,
  beta = 0.001,
  ncore = 50,
  method = c("EMMAP", "GIBBS"),
  sample.idx = seq(from = 502, to = 2500, by = 2),
  n.iter = 5000,
  thres = 1e-05
)
}
\arguments{
\item{bulk}{bulk RNA-seq data. Counts. Genes by samples.}

\item{phi}{cell type-specific gene expression references. A list of length of the number of cell types. Each element of the list is a genes by reference matrix.}

\item{alpha}{numeric. prior for cellular fraction Dirichlet distribution.}

\item{beta}{numeric. prior for reference mixing proportion Dirichlet distribution.}

\item{ncore}{the number of cores that will be used for computation.}

\item{method}{posterior inference technique that will be used.}

\item{sample.idx}{vector, index of Gibbs samples to be used for parameter estimation.}

\item{n.iter}{maximum iterations allowed in EM-MAP.}

\item{thres}{stopping criterion in EM-MAP.}
}
\value{
a list whose elements are the estimation results for each bulk sample
}
\description{
Estimate cell type fractions of bulk RNA-seq data using multiple references
}
